Service 로직 추가
