+++ routes/index.js
@@ -1,32 +1,11 @@
 const express = require('express');
 const router = express.Router();
-let Workspace = require('../models/index').models.workspace;
 
-let mailer = require('../auth/mail');
-
-router.get('/', async(req, res, next) => {
-  console.log(Workspace);
-  const data = await Workspace.findAll();
-  res.send({data});
-});
+const indexService = require('../service/indexService');
 
 router.post('/mail', async(req, res) => {
-  const { email }  = req.body;
-  const min = 111111;
-  const max = 999999;
-  const ranNum = Math.floor(Math.random() * (max - min - 1)) + min;
-
-  let emailParam = {
-    toEmail: email,     // 수신할 이메일
-
-    subject: `[slack_clone]인증 관련 메일 입니다`,   // 메일 제목
-
-    text: `오른쪽 숫자 6자리를 입력해주세요 : ${ranNum}`                // 메일 내용
-  };
-
-  mailer.sendGmail(emailParam);
-
-  res.status(200).send("성공");
+  const ranNum = await indexService.sendMail(req.body);
+  res.status(200).send(ranNum);
 });
 
 module.exports = router;
